+++ cpp/src/io/csv/csv_reader.cu
@@ -52,14 +52,12 @@
 
 #include "cudf.h"
 #include "utilities/error_utils.h"
+#include "utilities/type_dispatcher.hpp"
 
 #include "rmm/rmm.h"
 #include "rmm/thrust_rmm_allocator.h"
 #include "io/comp/io_uncomp.h"
 
-#include "NVStrings.h"
-
-constexpr int32_t HASH_SEED = 33;
 constexpr size_t max_chunk_bytes = 64*1024*1024; // 64MB
 
 using std::vector;
@@ -74,12 +72,7 @@ typedef struct raw_csv_ {
     char *				data;			// on-device: the raw unprocessed CSV data - loaded as a large char * array
     cu_recstart_t*		recStart;		// on-device: Starting position of the records.
 
-    char				delimiter;		// host: the delimiter
-    char				terminator;		// host: the line terminator
-
-    char				quotechar;		// host: the quote character
-    bool				keepquotes;		// host: indicates to keep the start and end quotechar
-    bool				doublequote;	// host: indicates to interpret two consecutive quotechar as a single
+    ParseOptions        opts;			// host: options to control parsing behavior
 
     long				num_bytes;		// host: the number of bytes in the data
     long				num_bits;		// host: the number of 64-bit bitmaps (different than valid)
@@ -92,9 +85,6 @@ typedef struct raw_csv_ {
     bool* 				h_parseCol;		// host   : array of booleans stating if column should be parsed in reading process: parseCol[x]=false means that the column x needs to be filtered out.
     bool* 				d_parseCol;		// device : array of booleans stating if column should be parsed in reading process: parseCol[x]=false means that the column x needs to be filtered out.
     long 				header_row;		// Row id of the header
-    bool				dayfirst;
-    char				decimal;
-    char				thousands;
 
 	gdf_size_type  		nrows;       	// number of rows of file to read. default is -1, and all rows are read in this case
 	gdf_size_type		skiprows;       // number of rows at the start of the file to skip, default is 0
@@ -114,19 +104,6 @@ typedef struct column_data_ {
 	unsigned long long countInt64;
 	gdf_size_type countNULL;
 } column_data_t;
-
-typedef struct parsing_opts_ {
-	char				delimiter;
-	char				terminator;
-	char				quotechar;
-	bool				keepquotes;
-	char				decimal;
-	char				thousands;
-	int32_t*			trueValues;
-	int32_t*			falseValues;
-	int32_t				trueValuesCount;
-	int32_t				falseValuesCount;
-} parsing_opts_t;
 
 using string_pair = std::pair<const char*,size_t>;
 
@@ -162,11 +139,13 @@ __global__ void storeRecordStart(char *data, size_t chunk_offset,
 	const char terminator, const char quotechar, bool include_first_row,
 	long num_bytes, long num_bits, cu_reccnt_t* num_records,
 	cu_recstart_t* recStart);
-__global__ void convertCsvToGdf(char *csv, const parsing_opts_t opts, gdf_size_type num_records, int num_columns, 
-	bool *parseCol, cu_recstart_t *recStart,gdf_dtype *dtype, void **gdf_data, gdf_valid_type **valid, 
-	string_pair **str_cols, bool dayfirst, unsigned long long *num_valid);
-__global__ void dataTypeDetection(char *raw_csv, const parsing_opts_t opts, gdf_size_type num_records, 
-	int  num_columns, bool  *parseCol, cu_recstart_t *recStart, column_data_t* d_columnData);
+__global__ void convertCsvToGdf(char *csv, const ParseOptions opts,
+	gdf_size_type num_records, int num_columns, bool *parseCol,
+	cu_recstart_t *recStart, gdf_dtype *dtype, void **gdf_data, gdf_valid_type **valid,
+	string_pair **str_cols, unsigned long long *num_valid);
+__global__ void dataTypeDetection(char *raw_csv, const ParseOptions opts,
+	gdf_size_type num_records, int num_columns, bool *parseCol,
+	cu_recstart_t *recStart, column_data_t* d_columnData);
 
 //
 //---------------CUDA Valid (8 blocks of 8-bits) Bitmap Kernels ---------------------------------------------
@@ -185,27 +164,6 @@ __inline__ __device__ void validAtomicOR(gdf_valid_type* address, gdf_valid_type
 __device__ void setBit(gdf_valid_type* address, int bit) {
 	gdf_valid_type bitMask[8] 		= {1, 2, 4, 8, 16, 32, 64, 128};
 	validAtomicOR(address, bitMask[bit]);
-}
-
-
-
-
-std::string stringType(gdf_dtype dt){
-
-	switch (dt){
-		case GDF_STRING: return std::string("str");
-		case GDF_DATE64: return std::string("date64");
-		case GDF_CATEGORY: return std::string("category");
-		case GDF_FLOAT64: return std::string("float64");
-		case GDF_INT8: return std::string("int8");
-		case GDF_INT16: return std::string("int16");
-		case GDF_INT32: return std::string("int32");
-		case GDF_INT64: return std::string("int64");
-		default:
-			return "long";
-	}
-
-
 }
 
 
@@ -289,37 +247,12 @@ gdf_error read_csv(csv_read_arg *args)
 	raw_csv->num_active_cols	= args->num_cols;
 	raw_csv->num_records		= 0;
 
-	if(args->delim_whitespace == true) {
-		raw_csv->delimiter = ' ';
-	} else {
-		raw_csv->delimiter = args->delimiter;
-	}
-
-	if(args->windowslinetermination) {
-		raw_csv->terminator = '\n';
-	} else {
-		raw_csv->terminator = args->lineterminator;
-	}
-
-	raw_csv->quotechar = args->quotechar;
-	if(raw_csv->quotechar != '\0') {
-		raw_csv->keepquotes = !args->quoting;
-		raw_csv->doublequote = args->doublequote;
-	} else {
-		raw_csv->keepquotes = true;
-		raw_csv->doublequote = false;
-	}
-
 	if (args->names == nullptr) {
 		raw_csv->header_row = args->header;
 	}
 	else{
 		raw_csv->header_row = -1;
 	}
-
-	raw_csv->dayfirst = args->dayfirst;
-	raw_csv->decimal = args->decimal;
-	raw_csv->thousands = args->thousands;
 
 	raw_csv->skiprows = args->skiprows;
 	raw_csv->skipfooter = args->skipfooter;
@@ -328,12 +261,32 @@ gdf_error read_csv(csv_read_arg *args)
 		++raw_csv->nrows;
 	}
 
-	if (raw_csv->decimal == raw_csv->delimiter)
-	{ 
+	if (args->delim_whitespace) {
+		raw_csv->opts.delimiter = ' ';
+	} else {
+		raw_csv->opts.delimiter = args->delimiter;
+	}
+	if (args->windowslinetermination) {
+		raw_csv->opts.terminator = '\n';
+	} else {
+		raw_csv->opts.terminator = args->lineterminator;
+	}
+	if (args->quotechar != '\0') {
+		raw_csv->opts.quotechar = args->quotechar;
+		raw_csv->opts.keepquotes = !args->quoting;
+		raw_csv->opts.doublequote = args->doublequote;
+	} else {
+		raw_csv->opts.quotechar = args->quotechar;
+		raw_csv->opts.keepquotes = true;
+		raw_csv->opts.doublequote = false;
+	}
+	raw_csv->opts.dayfirst = args->dayfirst;
+	raw_csv->opts.decimal = args->decimal;
+	raw_csv->opts.thousands = args->thousands;
+	if (raw_csv->opts.decimal == raw_csv->opts.delimiter) {
 		checkError(GDF_INVALID_API_CALL, "Decimal point cannot be the same as the delimiter");
 	}
-	if (raw_csv->thousands == raw_csv->delimiter)
-	{ 
+	if (raw_csv->opts.thousands == raw_csv->opts.delimiter) {
 		checkError(GDF_INVALID_API_CALL, "Thousands separator cannot be the same as the delimiter");
 	}
 
@@ -359,17 +312,21 @@ gdf_error read_csv(csv_read_arg *args)
 	if ((args->true_values != NULL) && (args->num_true_values > 0)) {
 		thrust::host_vector<int32_t> h_values(args->num_true_values);
 		for (int i = 0; i < args->num_true_values; ++i) {
-			h_values[i] = convertStrtoInt<int32_t>(args->true_values[i], 0, strlen(args->true_values[i]) - 1);
+			h_values[i] = convertStrToValue<int32_t>(args->true_values[i], 0, strlen(args->true_values[i]) - 1, raw_csv->opts);
 		}
 		raw_csv->d_trueValues = h_values;
 	}
 	if ((args->false_values != NULL) && (args->num_false_values > 0)) {
 		thrust::host_vector<int32_t> h_values(args->num_false_values);
 		for (int i = 0; i < args->num_false_values; ++i) {
-			h_values[i] = convertStrtoInt<int32_t>(args->false_values[i], 0, strlen(args->false_values[i]) - 1);
+			h_values[i] = convertStrToValue<int32_t>(args->false_values[i], 0, strlen(args->false_values[i]) - 1, raw_csv->opts);
 		}
 		raw_csv->d_falseValues = h_values;
 	}
+	raw_csv->opts.trueValues       = raw_csv->d_trueValues.data().get();
+	raw_csv->opts.trueValuesCount  = raw_csv->d_trueValues.size();
+	raw_csv->opts.falseValues      = raw_csv->d_falseValues.data().get();
+	raw_csv->opts.falseValuesCount = raw_csv->d_falseValues.size();
 
 	//-----------------------------------------------------------------------------
 	// memory map in the data
@@ -437,7 +394,7 @@ gdf_error read_csv(csv_read_arg *args)
 	assert(h_uncomp_data != nullptr);
 	assert(h_uncomp_size != 0);
 
-	error = launch_countRecords(h_uncomp_data, h_uncomp_size, raw_csv->terminator, raw_csv->quotechar, raw_csv->num_records);
+	error = launch_countRecords(h_uncomp_data, h_uncomp_size, raw_csv->opts.terminator, raw_csv->opts.quotechar, raw_csv->num_records);
 	if (error != GDF_SUCCESS) {
 			return error;
 	}
@@ -465,7 +422,7 @@ gdf_error read_csv(csv_read_arg *args)
 	// the records of both, and then filtering out the records that is a quotechar
 	// or a linetermination within a quotechar pair. The future major refactoring
 	// of csv_reader and its kernels will probably use a different tactic.
-	if (raw_csv->quotechar != '\0') {
+	if (raw_csv->opts.quotechar != '\0') {
 		vector<cu_recstart_t> h_rec_starts(raw_csv->num_records + 1);
 		const size_t rec_start_size = sizeof(cu_recstart_t) * (h_rec_starts.size());
 		CUDA_TRY( cudaMemcpy(h_rec_starts.data(), raw_csv->recStart, rec_start_size, cudaMemcpyDeviceToHost) );
@@ -474,7 +431,7 @@ gdf_error read_csv(csv_read_arg *args)
 
 		bool quotation = false;
 		for (gdf_size_type i = 1; i < raw_csv->num_records; ++i) {
-			if (h_uncomp_data[h_rec_starts[i] - 1] == raw_csv->quotechar) {
+			if (h_uncomp_data[h_rec_starts[i] - 1] == raw_csv->opts.quotechar) {
 				quotation = !quotation;
 				h_rec_starts[i] = raw_csv->num_bytes;
 				recCount--;
@@ -778,7 +735,7 @@ gdf_error read_csv(csv_read_arg *args)
 		h_dtypes[col] 	= gdf->dtype;
 		h_data[col] 	= gdf->data;
 		h_valid[col] 	= gdf->valid;	
-        }
+    }
 
 	CUDA_TRY( cudaMemcpy(d_dtypes,h_dtypes, sizeof(gdf_dtype) * (raw_csv->num_active_cols), cudaMemcpyHostToDevice));
 	CUDA_TRY( cudaMemcpy(d_data,h_data, sizeof(void*) * (raw_csv->num_active_cols), cudaMemcpyHostToDevice));
@@ -805,11 +762,11 @@ gdf_error read_csv(csv_read_arg *args)
 				continue;
 
 			NVStrings* const stringCol = NVStrings::create_from_index(h_str_cols[stringColCount],size_t(raw_csv->num_records));
-			if ((raw_csv->quotechar != '\0') && (raw_csv->doublequote==true)) {
+			if ((raw_csv->opts.quotechar != '\0') && (raw_csv->opts.doublequote==true)) {
 				// In PANDAS, default of enabling doublequote for two consecutive
 				// quotechar in quote fields results in reduction to single
-				const string quotechar(1, raw_csv->quotechar);
-				const string doublequotechar(2, raw_csv->quotechar);
+				const string quotechar(1, raw_csv->opts.quotechar);
+				const string doublequotechar(2, raw_csv->opts.quotechar);
 				gdf->data = stringCol->replace(doublequotechar.c_str(), quotechar.c_str());
 				NVStrings::destroy(stringCol);
 			}
@@ -1020,63 +977,35 @@ gdf_error uploadDataToDevice(const char* h_uncomp_data, size_t h_uncomp_size, ra
 	return GDF_SUCCESS;
 }
 
-/*
- * For each of the gdf_cvolumns, create the on-device space.  the on-host fields should already be filled in
- */
-gdf_error allocateGdfDataSpace(gdf_column *gdf) {
 
-	long N = gdf->size;
-	long num_bitmaps = (N + 31) / 8;			// 8 bytes per bitmap
+/**---------------------------------------------------------------------------*
+ * @brief Allocates memory for a column's parsed output and its validity bitmap
+ *
+ * Memory for column data is simply based upon number of rows and the size of
+ * the output data type, regardless of actual validity of the row element.
+ *
+ * @param[in,out] col The column whose memory will be allocated
+ *
+ * @return gdf_error GDF_SUCCESS upon completion
+ *---------------------------------------------------------------------------**/
+gdf_error allocateGdfDataSpace(gdf_column *col) {
+  // TODO: We should not need to allocate space if there is nothing to parse
+  // Need to debug/refactor the code to eliminate this requirement
+  const auto num_rows = std::max(col->size, 1);
+  const auto num_masks = gdf_get_num_chars_bitmask(num_rows);
 
-	//--- allocate space for the valid bitmaps
-	RMM_TRY( RMM_ALLOC((void**)&gdf->valid, (sizeof(gdf_valid_type) * num_bitmaps), 0) );
-	CUDA_TRY(cudaMemset(gdf->valid, 0, (sizeof(gdf_valid_type) 	* num_bitmaps)) );
+  RMM_TRY(RMM_ALLOC(&col->valid, sizeof(gdf_valid_type) * num_masks, 0));
+  CUDA_TRY(cudaMemset(col->valid, 0, sizeof(gdf_valid_type) * num_masks));
 
-	int elementSize=0;
-	//--- Allocate space for the data
-	switch(gdf->dtype) {
-		case gdf_dtype::GDF_INT8:
-			elementSize = sizeof(int8_t);
-			break;
-		case gdf_dtype::GDF_INT16:
-			elementSize = sizeof(int16_t);
-			break;
-		case gdf_dtype::GDF_INT32:
-			elementSize = sizeof(int32_t);
-			break;
-		case gdf_dtype::GDF_INT64:
-			elementSize = sizeof(int64_t);
-			break;
-		case gdf_dtype::GDF_FLOAT32:
-			elementSize = sizeof(float);
-			break;
-		case gdf_dtype::GDF_FLOAT64:
-			elementSize = sizeof(double);
-			break;
-		case gdf_dtype::GDF_DATE32:
-			elementSize = sizeof(gdf_date32);
-			break;
-		case gdf_dtype::GDF_DATE64:
-			elementSize = sizeof(gdf_date64);
-			break;
-		case gdf_dtype::GDF_TIMESTAMP:
-			elementSize = sizeof(int64_t);
-			break;
-		case gdf_dtype::GDF_CATEGORY:
-			elementSize = sizeof(gdf_category);
-			break;
-		case gdf_dtype::GDF_STRING:
-			return gdf_error::GDF_SUCCESS;
-			// Memory for gdf->data allocated by string class eventually
-		default:
-			return GDF_UNSUPPORTED_DTYPE;
-	}
-	
-	RMM_TRY( RMM_ALLOC((void**)&gdf->data, elementSize * N, 0) );
+  if (col->dtype != gdf_dtype::GDF_STRING) {
+    int column_byte_width = 0;
+    checkError(get_column_byte_width(col, &column_byte_width),
+               "Could not get column width using data type");
+    RMM_TRY(RMM_ALLOC(&col->data, num_rows * column_byte_width, 0));
+  }
 
-	return gdf_error::GDF_SUCCESS;
+  return GDF_SUCCESS;
 }
-
 
 //----------------------------------------------------------------------------------------------------------------
 //				CUDA Kernels
@@ -1243,7 +1172,7 @@ gdf_error launch_storeRecordStart(const char* h_data, size_t h_size, bool includ
 
 		const int gridSize = (chunk_bits + blockSize - 1) / blockSize;
 		storeRecordStart <<< gridSize, blockSize >>> (
-			d_chunk, chunk_offset, csvData->terminator, csvData->quotechar, cu_include_first_row,
+			d_chunk, chunk_offset, csvData->opts.terminator, csvData->opts.quotechar, cu_include_first_row,
 			chunk_bytes, chunk_bits, d_num_records,
 			csvData->recStart
 		);
@@ -1322,9 +1251,17 @@ __global__ void storeRecordStart(char *data, size_t chunk_offset,
 }
 
 
-//----------------------------------------------------------------------------------------------------------------
-
-
+/**---------------------------------------------------------------------------*
+ * @brief Helper function to setup and launch CSV parsing CUDA kernel.
+ * 
+ * @param[in,out] raw_csv The metadata for the CSV data
+ * @param[out] gdf The output column data
+ * @param[out] valid The bitmaps indicating whether column fields are valid
+ * @param[out] str_cols The start/end offsets for string data types
+ * @param[out] num_valid The numbers of valid fields in columns
+ *
+ * @return gdf_error GDF_SUCCESS upon completion
+ *---------------------------------------------------------------------------**/
 gdf_error launch_dataConvertColumns(raw_csv_t *raw_csv, void **gdf, gdf_valid_type** valid, gdf_dtype* d_dtypes,string_pair **str_cols, unsigned long long *num_valid) {
 
 	int blockSize;		// suggested thread count to use
@@ -1334,27 +1271,15 @@ gdf_error launch_dataConvertColumns(raw_csv_t *raw_csv, void **gdf, gdf_valid_ty
 	// Calculate actual block count to use based on records count
 	int gridSize = (raw_csv->num_records + blockSize - 1) / blockSize;
 
-	parsing_opts_t opts;
-	opts.delimiter			= raw_csv->delimiter;
-	opts.terminator			= raw_csv->terminator;
-	opts.quotechar			= raw_csv->quotechar;
-	opts.keepquotes			= raw_csv->keepquotes;
-	opts.decimal			= raw_csv->decimal;
-	opts.thousands			= raw_csv->thousands;
-	opts.trueValues			= thrust::raw_pointer_cast(raw_csv->d_trueValues.data());
-	opts.trueValuesCount		= raw_csv->d_trueValues.size();
-	opts.falseValues		= thrust::raw_pointer_cast(raw_csv->d_falseValues.data());
-	opts.falseValuesCount		= raw_csv->d_falseValues.size();
-
 	auto first_data_rec_start = raw_csv->recStart;
 	if (raw_csv->header_row >= 0) {
 		// skip the header row if present
 		++first_data_rec_start;
 	}
 
-	convertCsvToGdf <<< gridSize, blockSize >>>(
+	convertCsvToGdf <<< gridSize, blockSize >>> (
 		raw_csv->data,
-		opts,
+		raw_csv->opts,
 		raw_csv->num_records,
 		raw_csv->num_actual_cols,
 		raw_csv->d_parseCol,
@@ -1363,7 +1288,6 @@ gdf_error launch_dataConvertColumns(raw_csv_t *raw_csv, void **gdf, gdf_valid_ty
 		gdf,
 		valid,
 		str_cols,
-		raw_csv->dayfirst,
 		num_valid
 	);
 
@@ -1371,25 +1295,81 @@ gdf_error launch_dataConvertColumns(raw_csv_t *raw_csv, void **gdf, gdf_valid_ty
 	return GDF_SUCCESS;
 }
 
+/**---------------------------------------------------------------------------*
+ * @brief Functor for converting CSV data to cuDF data type value.
+ *---------------------------------------------------------------------------**/
+struct ConvertFunctor {
+  /**---------------------------------------------------------------------------*
+   * @brief Template specialization for operator() that handles integer types
+   * that additionally checks whether the parsed data value should be overridden
+   * with user-specified true/false matches.
+   *
+   * It is handled here rather than within convertStrToValue() as that function
+   * is already used to construct the true/false match list from user-provided
+   * strings at the start of parsing.
+   *---------------------------------------------------------------------------**/
+  template <typename T,
+            typename std::enable_if_t<std::is_integral<T>::value> * = nullptr>
+  __host__ __device__ __forceinline__ void operator()(
+      const char *csvData, void *gdfColumnData, long rowIndex, long start,
+      long end, const ParseOptions &opts) {
+    T &value{static_cast<T *>(gdfColumnData)[rowIndex]};
+    value = convertStrToValue<T>(csvData, start, end, opts);
 
-/*
- * Data is processed in one row\record at a time - so the number of total threads (tid) is equal to the number of rows.
+    // Check for user-specified true/false values where the output is
+    // replaced with 1/0 respectively
+    if (isBooleanValue(value, opts.trueValues, opts.trueValuesCount)) {
+      value = 1;
+    } else if (isBooleanValue(value, opts.falseValues, opts.falseValuesCount)) {
+      value = 0;
+    }
+  }
+
+  /**---------------------------------------------------------------------------*
+   * @brief Default template operator() dispatch specialization all data types
+   * (including wrapper types) that is not covered by integral specialization.
+   *---------------------------------------------------------------------------**/
+  template <typename T,
+            typename std::enable_if_t<!std::is_integral<T>::value> * = nullptr>
+  __host__ __device__ __forceinline__ void operator()(
+      const char *csvData, void *gdfColumnData, long rowIndex, long start,
+      long end, const ParseOptions &opts) {
+    T &value{static_cast<T *>(gdfColumnData)[rowIndex]};
+    value = convertStrToValue<T>(csvData, start, end, opts);
+  }
+};
+
+/**---------------------------------------------------------------------------*
+ * @brief CUDA kernel that parses and converts CSV data into cuDF column data.
+ * 
+ * Data is processed one record at a time
  *
- */
-__global__ void convertCsvToGdf(
-		char 			*raw_csv,
-		const parsing_opts_t	 	opts,
-		gdf_size_type  num_records,
-		int  			num_columns,
-		bool  			*parseCol,
-		cu_recstart_t 			*recStart,
-		gdf_dtype 		*dtype,
-		void			**gdf_data,
-		gdf_valid_type 	**valid,
-		string_pair		**str_cols,
-		bool			dayfirst,
-		unsigned long long			*num_valid
-		)
+ * @param[in] raw_csv The entire CSV data to read
+ * @param[in] opts A set of parsing options
+ * @param[in] num_records The number of lines/rows of CSV data
+ * @param[in] num_columns The number of columns of CSV data
+ * @param[in] parseCol Whether to parse or skip a column
+ * @param[in] recStart The start the CSV data of interest
+ * @param[in] dtype The data type of the column
+ * @param[out] gdf_data The output column data
+ * @param[out] valid The bitmaps indicating whether column fields are valid
+ * @param[out] str_cols The start/end offsets for string data types
+ * @param[out] num_valid The numbers of valid fields in columns
+ *
+ * @return gdf_error GDF_SUCCESS upon completion
+ *---------------------------------------------------------------------------**/
+__global__
+void convertCsvToGdf(char *raw_csv,
+                     const ParseOptions opts,
+                     gdf_size_type num_records,
+                     int num_columns,
+                     bool *parseCol,
+                     cu_recstart_t *recStart,
+                     gdf_dtype *dtype,
+                     void **gdf_data,
+                     gdf_valid_type **valid,
+                     string_pair **str_cols,
+                     unsigned long long *num_valid)
 {
 	// thread IDs range per block, so also need the block id
 	long	rec_id  = threadIdx.x + (blockDim.x * blockIdx.x);		// this is entry into the field array - tid is an elements within the num_entries array
@@ -1445,108 +1425,24 @@ __global__ void convertCsvToGdf(
 				adjustForWhitespaceAndQuotes(raw_csv, &start, &tempPos, opts.quotechar);
 			}
 
-			if(start<=(tempPos)) { // Empty strings are not legal values
+			if(start<=(tempPos)) { // Empty fields are not legal values
 
-				switch(dtype[actual_col]) {
-					case gdf_dtype::GDF_INT8:
-					{
-						int8_t *gdf_out = (int8_t *)gdf_data[actual_col];
-						gdf_out[rec_id] = convertStrtoInt<int8_t>(raw_csv, start, tempPos, opts.thousands);
-
-						if(isBooleanValue(gdf_out[rec_id], opts.trueValues, opts.trueValuesCount)==true){
-							gdf_out[rec_id] = 1;
-						}else if(isBooleanValue(gdf_out[rec_id], opts.falseValues, opts.falseValuesCount)==true){
-							gdf_out[rec_id] = 0;
+				// Type dispatcher does not handle GDF_STRINGS
+				if (dtype[actual_col] == gdf_dtype::GDF_STRING) {
+					long end = pos;
+					if(opts.keepquotes==false){
+						if((raw_csv[start] == opts.quotechar) && (raw_csv[end-1] == opts.quotechar)){
+							start++;
+							end--;
 						}
 					}
-						break;
-					case gdf_dtype::GDF_INT16: {
-						int16_t *gdf_out = (int16_t *)gdf_data[actual_col];
-						gdf_out[rec_id] = convertStrtoInt<int16_t>(raw_csv, start, tempPos, opts.thousands);
-
-						if(isBooleanValue(gdf_out[rec_id], opts.trueValues, opts.trueValuesCount)==true){
-							gdf_out[rec_id] = 1;
-						}else if(isBooleanValue(gdf_out[rec_id], opts.falseValues, opts.falseValuesCount)==true){
-							gdf_out[rec_id] = 0;
-						}
-					}
-						break;
-					case gdf_dtype::GDF_INT32:
-					{
-						int32_t *gdf_out = (int32_t *)gdf_data[actual_col];
-						gdf_out[rec_id] = convertStrtoInt<int32_t>(raw_csv, start, tempPos, opts.thousands);
-
-						if(isBooleanValue(gdf_out[rec_id], opts.trueValues, opts.trueValuesCount)==true){
-							gdf_out[rec_id] = 1;
-						}else if(isBooleanValue(gdf_out[rec_id], opts.falseValues, opts.falseValuesCount)==true){
-							gdf_out[rec_id] = 0;
-						}
-					}
-						break;
-					case gdf_dtype::GDF_INT64:
-					{
-						int64_t *gdf_out = (int64_t *)gdf_data[actual_col];
-						gdf_out[rec_id] = convertStrtoInt<int64_t>(raw_csv, start, tempPos, opts.thousands);
-
-						if(isBooleanValue(gdf_out[rec_id], opts.trueValues, opts.trueValuesCount)==true){
-							gdf_out[rec_id] = 1;
-						}else if(isBooleanValue(gdf_out[rec_id], opts.falseValues, opts.falseValuesCount)==true){
-							gdf_out[rec_id] = 0;
-						}
-					}
-						break;
-					case gdf_dtype::GDF_FLOAT32:
-					{
-						float *gdf_out = (float *)gdf_data[actual_col];
-						gdf_out[rec_id] = convertStrtoFloat<float>(raw_csv, start, tempPos, opts.decimal, opts.thousands);
-					}
-						break;
-					case gdf_dtype::GDF_FLOAT64:
-					{
-						double *gdf_out = (double *)gdf_data[actual_col];
-						gdf_out[rec_id] = convertStrtoFloat<double>(raw_csv, start, tempPos, opts.decimal, opts.thousands);
-					}
-						break;
-					case gdf_dtype::GDF_DATE32:
-					{
-						gdf_date32 *gdf_out = (gdf_date32 *)gdf_data[actual_col];
-						gdf_out[rec_id] = parseDateFormat(raw_csv, start, tempPos, dayfirst);
-					}
-						break;
-					case gdf_dtype::GDF_DATE64:
-					{
-						gdf_date64 *gdf_out = (gdf_date64 *)gdf_data[actual_col];
-						gdf_out[rec_id] = parseDateTimeFormat(raw_csv, start, tempPos, dayfirst);
-					}
-						break;
-					case gdf_dtype::GDF_TIMESTAMP:
-					{
-						int64_t *gdf_out = (int64_t *)gdf_data[actual_col];
-						gdf_out[rec_id] = convertStrtoInt<int64_t>(raw_csv, start, tempPos, opts.thousands);
-					}
-					break;
-					case gdf_dtype::GDF_CATEGORY:
-					{
-						gdf_category *gdf_out = (gdf_category *)gdf_data[actual_col];
-						gdf_out[rec_id] = convertStrtoHash(raw_csv, start, pos, HASH_SEED);
-					}
-						break;
-					case gdf_dtype::GDF_STRING:
-					{
-						long end = pos;
-						if(opts.keepquotes==false){
-							if((raw_csv[start] == opts.quotechar) && (raw_csv[end-1] == opts.quotechar)){
-								start++;
-								end--;
-							}
-						}
-						str_cols[stringCol][rec_id].first	= raw_csv+start;
-						str_cols[stringCol][rec_id].second	= size_t(end-start);
-						stringCol++;
-					}
-						break;
-					default:
-						break;
+					str_cols[stringCol][rec_id].first	= raw_csv+start;
+					str_cols[stringCol][rec_id].second	= size_t(end-start);
+					stringCol++;
+				} else {
+					cudf::type_dispatcher(
+						dtype[actual_col], ConvertFunctor{}, raw_csv,
+						gdf_data[actual_col], rec_id, start, tempPos, opts);
 				}
 
 				// set the valid bitmap - all bits were set to 0 to start
@@ -1570,14 +1466,16 @@ __global__ void convertCsvToGdf(
 	}
 }
 
-
-
-//----------------------------------------------------------------------------------------------------------------
-
-
-gdf_error launch_dataTypeDetection(
-	raw_csv_t * raw_csv,
-	column_data_t* d_columnData) 
+/**---------------------------------------------------------------------------*
+ * @brief Helper function to setup and launch CSV data type detect CUDA kernel.
+ * 
+ * @param[in] raw_csv The metadata for the CSV data
+ * @param[out] d_columnData The count for each column data type
+ *
+ * @return gdf_error GDF_SUCCESS upon completion
+ *---------------------------------------------------------------------------**/
+gdf_error launch_dataTypeDetection(raw_csv_t *raw_csv,
+                                   column_data_t *d_columnData)
 {
 	int blockSize;		// suggested thread count to use
 	int minGridSize;	// minimum block count required
@@ -1586,25 +1484,15 @@ gdf_error launch_dataTypeDetection(
 	// Calculate actual block count to use based on records count
 	int gridSize = (raw_csv->num_records + blockSize - 1) / blockSize;
 
-	parsing_opts_t opts;
-	opts.delimiter			= raw_csv->delimiter;
-	opts.terminator			= raw_csv->terminator;
-	opts.quotechar			= raw_csv->quotechar;
-	opts.keepquotes			= raw_csv->keepquotes;
-	opts.trueValues			= thrust::raw_pointer_cast(raw_csv->d_trueValues.data());
-	opts.trueValuesCount		= raw_csv->d_trueValues.size();
-	opts.falseValues		= thrust::raw_pointer_cast(raw_csv->d_falseValues.data());
-	opts.falseValuesCount		= raw_csv->d_falseValues.size();
-
 	auto first_data_rec_start = raw_csv->recStart;
 	if (raw_csv->header_row >= 0) {
 		// skip the header row if present
 		++first_data_rec_start;
 	}
 
-	dataTypeDetection <<< gridSize, blockSize >>>(
+	dataTypeDetection <<< gridSize, blockSize >>> (
 		raw_csv->data,
-		opts,
+		raw_csv->opts,
 		raw_csv->num_records,
 		raw_csv->num_actual_cols,
 		raw_csv->d_parseCol,
@@ -1616,17 +1504,30 @@ gdf_error launch_dataTypeDetection(
 	return GDF_SUCCESS;
 }
 
-/*
- */
-__global__ void dataTypeDetection(
-		char 			*raw_csv,
-		const parsing_opts_t			opts,
-		gdf_size_type  			num_records,
-		int  			num_columns,
-		bool  			*parseCol,
-		cu_recstart_t 			*recStart,
-		column_data_t* d_columnData
-		)
+/**---------------------------------------------------------------------------*
+ * @brief CUDA kernel that parses and converts CSV data into cuDF column data.
+ *
+ * Data is processed in one row/record at a time, so the number of total
+ * threads (tid) is equal to the number of rows.
+ *
+ * @Param[in] raw_csv The entire CSV data to read
+ * @Param[in] opts A set of parsing options
+ * @Param[in] num_records The number of lines/rows of CSV data
+ * @Param[in] num_columns The number of columns of CSV data
+ * @Param[in] parseCol Whether to parse or skip a column
+ * @Param[in] recStart The start the CSV data of interest
+ * @Param[out] d_columnData The count for each column data type
+ *
+ * @Returns GDF_SUCCESS upon successful computation
+ *---------------------------------------------------------------------------**/
+__global__
+void dataTypeDetection(char *raw_csv,
+                       const ParseOptions opts,
+                       gdf_size_type num_records,
+                       int num_columns,
+                       bool *parseCol,
+                       cu_recstart_t *recStart,
+                       column_data_t *d_columnData)
 {
 	// thread IDs range per block, so also need the block id
 	long	rec_id  = threadIdx.x + (blockDim.x * blockIdx.x);		// this is entry into the field array - tid is an elements within the num_entries array
@@ -1731,7 +1632,7 @@ __global__ void dataTypeDetection(
 			else if(countNumber==(strLen) || ( strLen>1 && countNumber==(strLen-1) && raw_csv[start]=='-') ){
 				// Checking to see if we the integer value requires 8,16,32,64 bits.
 				// This will allow us to allocate the exact amount of memory.
-				int64_t value = convertStrtoInt<int64_t>(raw_csv, start, tempPos, opts.thousands);
+				const auto value = convertStrToValue<int64_t>(raw_csv, start, tempPos, opts);
 
 				if (isBooleanValue<int32_t>(value, opts.trueValues, opts.trueValuesCount) ||
 					isBooleanValue<int32_t>(value, opts.falseValues, opts.falseValuesCount)){