+++ CHANGELOG.md
@@ -8,11 +8,13 @@
 ## Improvements
 
 - PR #730 Improve performance of `gdf_table` constructor
+- PR #532 CSV Reader: Use type dispatcher instead of switch block
 
 ## Bug Fixes
 
 - PR #821 Fix flake8 issues revealed by flake8 update
 - PR #808 Resolved renamed `d_columns_valids` variable name
+- PR #780 CSV Reader: Fix scientific notation parsing and null values for empty quotes
 
 
 # cuDF 0.5.0 (28 Jan 2019)